Fixed a crashed due to unexpected data type from Get-ChildItem
The original design also assumed that the output would be an ArrayList, and then the rest would follow-through as a multiple-instance.  However, it is possible to have a single item as - FileInfo instead of the expected ArrayList.  As such, this would ultimately cause problems due to this data-type mixing.  To get around this, I will use a dynamic variable and convert the results as an ArrayList.

Essentially, if the user had multiple or no files within the designated directory - no crash would occur.  However, if only ONE file was present - than a crash would occur.