Created a new function, __SetupEnvironment()
This function will help to reduce immediate re-draws of the terminal environment setup.  This is necessary as other sub-functions within the same class, offers the exact same information - yet the first is cleared to allow the second to be drawn.  This causes the terminal to visually act strange to the user as information seems 'hidden'.  Though, behind the hood - nothing is different, but this could cause confusion as with the re-draw.  With this function, we can control what function gets to use these instructions to help eliminate the re-drawing.