Added an overload BrowseFile()
Changed the initial function, BrowseFile(), to accept one additional argument for 'Initial Directory' as a string.  This defines where the File Browser opens to the specified directory requested.

The Overload function only exists due to the previous implementations requiring the User's Home Directory.  Because PowerShell does not yet have a default argument, this Overload will act as such for the time being.