Added an overload function for BrowseDirectory()
Changed the initial function, BrowseDirectory(), to accept one additional argument for 'Initial Directory' as a string. This defines where the Directory Browser opens to the specified directory requested.

The Overload function only exists due to the previous implementations requiring the User's Home Directory. Because PowerShell does not yet have a default argument, this Overload will act as such for the time being.