Revised __ReadMetaFile() such that it now uses ProjectMetaData
Changed the signature of __ReadMetaFile() to use the Project's Meta Data structure instead of individual variable references.  This change serves two purposes: First, to minimize the changes needed after further changes are made to the Meta data file - such as more information in which requires more arguments, and Second to avoid having to change all functions calls to __ReadMetaFile().