[WIP] Uninstallation
A very much WIP of the uninstallation functionality.  Presently, it doesn't work.  However, it is able to allow the user to select what they want to remove, alert the user exactly what the project is being considered to be removed, and confirms with the user.
The actual removal does not yet take place at this moment.  More work is needed.