Check to see if the Project's Dir. is empty
If the directory is empty, thus the $dynamicVariable will be $NULL, the __GetInstalledProjects() would try to add a null instance into an arraylist - throwing an error.  Now, it will check if the value is $null before attempting to use it.   If the value is null, this function will just return $true.