Added DirName, renamed FileName, FilePath
Added a new member variable __directoryName, which is intended to hold the parent directory that is containing the meta file.  Getters and Setters applied as GetDirectoryName() and SetDirectoryName().

Renamed __fileName to __metaFileName, Getters and Setters updated to SetMetaFileName() and GetMetaFileName()
Renamed __filePath to __metaFilePath.  Getters and Setters updated to SetMetaFilePath() and GetMetaFilePath()