Uninstalling Projects uses a Table instead of Directory Browser
Changed the way the user can removed PSCAT projects from the environment.  Instead of using the DirectoryBrowser(), in which the user can accidentally navigate away from the projects install path, the user will now use the table populated with projects per-row.

This should be a much smoother experience for the user.