Added ReadTextFile()
This function will read the file provided, using the Get-Content CMDLet, and then return the results as an ArrayList.

This function was necessary as the Get-Content CMDLet returns one of the two datatypes:
 - String = when only one string was found in the textfile
 - ArrayList = When two or more strings was found in the textfile.

This, in the end, will help to reduce a bug that could happen when only one string was found - but required needing an ArrayList datatype.  This issue would cause the program to behave incorrectly.  Lets keep things /CONSISTENT/, in POSH.