Corrected initialization of $getModuleInfo
Corrected an issue where the initialization of $getModuleInfo was not correctly performed.  Instead of using the new() statement, we can just assign it to $NULL.  Otherwise, using this function will result in a crash as the new() statement is not available or requires arguments that is not yet known early in the function's lifespan.