Revised how we obtain data from a POSH Module
First off, using the Get-Module CMDlet, we must use the '-ListAvailable' option.  Without this, we do not retrieve any data....

Second, encapsulated the Get-Module CMDlet into a try{} if incase: a general error happens, OR if more than one POSH Module was detected but cannot assign it to $getModuleInfo as it is a single instance.  This requires a bit more testing.

Thirdly, and lastly, if an error was detected - than we can safely abort without causing the PSCAT to crash.

Oh, and added debug logging functionality so we can inspect what is happening under the bonnet when necessary.