Revised the flow of DetectPowerShellModule()
Revised the flow of the DetectPowerShellModule() function, such that we now use the try\catch block for when the module was not detected.  This method should be a bit more cleaner, as well as how data is now recorded.

Further, spent extra time revising the documentation within the function statements as well as documenting the function itself.

#88