Make supports caching program contents
The Make script is now capable of caching all of the sub-script contents into a string, once caching is completed - the Make script will then write all of the cached contents to the output file.  This feature is toggable using the global variable, $CACHEPROGRAMCONTENT.  This variable is set $true by default.

NOTE: For some - odd - reason, caching the data is a much heavier performance impact as opposed to writing to the disk per-iteration.  But thus far, it should solve the chief problem with Add-Content CMDlet.

#89