Added try\catch when caching data to global var.
Added a try\catch when caching data into the mutable global variable, $CACHEDPROGRAMCONTENTS.  If incase an error where to ever happen while performing this operation, we will be able to support it and show the exception to the user.

It /might/ be possible to hit a Common Language Runtime memory limit; if incase that were to ever happen - this will detect it for us.

#89