Created a new function, GetPowerShellModuleName()
Added a new function, GetPowerShellModuleName(), which will provide the PowerShell Module Name to non-DefaultCompress class, but also allowing a safer way to obtain the Module Name within the member functions\variables.  With that said, I had also updated the DetectCompressModule() to use the GetPowerShellModuleName() instead of using the private member variable directly -- if incase debugging or behaviorial changes are needed.