Revised Verbose Logging when POSH Module not avail.
I do not think it is possible to obtain the dotNET Core anymore; I think I wrote this in the early days when dotNET Core was not entirely bundled within POSH Core?  I could had swore you had to install both back in the day, but not really needed these days?  Hard for me to confirm 😕

#94