Removed Global Var, $_DEBUGLOGGING_
Logging through out this program is mandatory.  The reason for this being a necessity - is only for debugging purposes only.  IF something were to break, how can one correct it?  What arguments were used?  Many other questions like that.

If we trust that this program will never crash\bug-out or do not care about debugging it, then we can remove all debugging functionality.

#97